backend/requirements.txt
asgiref==3.2.7            # via django
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
django==3.0.5             # via -r requirements.in, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
idna==2.9                 # via requests
pyparsing==2.4.7          # via packaging
pytest-django==3.9.0      # via -r requirements.in
pytz==2019.3              # via django
requests==2.23.0          # via -r requirements.in
six==1.14.0               # via packaging
sqlparse==0.3.1           # via django
urllib3==1.25.9           # via requests
zipp==3.1.0               # via importlib-metadata
django_filter
djangorestframework_simplejwt
djoser
markdown
gunicorn==20.1.0
psycopg2>=2.8,<2.9
Pillow
drf_extra_fields